refactor: 명언 앱 코드를 객체지향적으로 리팩토링
애플리케이션의 유지보수성, 확장성, 테스트 용이성을 향상시키기 위해
코드를 객체지향 원칙에 따라 재구조화

주요 변경사항:
- App 클래스를 생성하여 앱의 전반적인 실행 흐름 및 명령어 처리 로직 위임
- QuoteRepository 클래스를 분리하여 명언 데이터 관리(CRUD) 전담
- Main 클래스를 App 객체 실행만을 위한 진입점 역할로 최소화
- Quote 클래스는 데이터 모델 역할 유지

캡슐화 및 의존성 주입 개념을 도입하여 코드의 결합도를 낮추고 유연성을 높였습니다.
